requirements.txt
streamlit>=1.35
gspread>=6.0
google-auth>=2.0
pandas>=2.0
plotly>=5.0
python-dotenv>=1.0
